Starting to stub out a data_request_spec struct
