+++ src/cli.rs
@@ -4,26 +4,8 @@ use clap::{Args, Parser, Subcommand};
 use serde::{Deserialize, Serialize};
 use serde_json;
 use strum_macros::EnumString;
+use popgetter::data_request_spec::BBox;
 
-#[derive(Clone, Debug)]
-pub struct BBox(pub [f64; 4]);
-
-impl FromStr for BBox {
-    type Err = &'static str;
-    fn from_str(value: &str) -> Result<Self, Self::Err> {
-        let parts: Vec<f64> = value
-            .split(',')
-            .map(|s| s.trim().parse::<f64>().map_err(|_| "Failed to parse bbox"))
-            .collect::<Result<Vec<_>, _>>()?;
-
-        if parts.len() != 4 {
-            return Err("Bounding boxes need to have 4 coords");
-        }
-        let mut bbox = [0.0; 4];
-        bbox.copy_from_slice(&parts);
-        Ok(BBox(bbox))
-    }
-}
 
 #[derive(Clone,Debug,Deserialize,Serialize,EnumString,PartialEq, Eq)]
 #[strum(ascii_case_insensitive)]
@@ -84,22 +66,6 @@ pub enum Commands {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn bbox_should_parse_if_correct() {
-        let bbox = BBox::from_str("0.0,1.0,2.0,3.0");
-        assert!(bbox.is_ok(), "A four coord bbox should parse");
-    }
-
-    #[test]
-    fn bbox_should_not_parse_if_incorrect() {
-        let bbox = BBox::from_str("0.0,1.0,2.0");
-        assert!(bbox.is_err(), "A string with fewer than 4 coords should parse");
-        let bbox = BBox::from_str("0.0,1.0,2.0,3.0,4.0");
-        assert!(bbox.is_err(), "A string with fewer than 5 coords should parse");
-        let bbox = BBox::from_str("0.0sdfsd,1.0,2.0");
-        assert!(bbox.is_err(), "A string with letters shouldn't parse");
-    }
 
     #[test]
     fn output_type_should_deserialize_properly(){
