+++ src/lib.rs
@@ -1,1 +1,2 @@
+pub mod data_request_spec;
 pub mod metadata;
